# tests/Fixtures/openapi.v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "MolMeDB: Molecules on Membranes Database",
        "version": "1.0.0",
        "description": "A database of molecules, membranes and their interactions (permeability, partitioning, binding), curated for membrane-transport research."
    },
    "servers": [
        {
            "url": "https://molmedb.upol.cz/api/v1",
            "description": "MolMeDB"
        }
    ],
    "paths": {
...
        "/membranes/{membrane}/interactions": {
            "get": {
                "operationId": "membrane.interactions",
                "tags": [
                    "Membrane"
                ],
                "parameters": [
                    {
                        "name": "membrane",
                        "in": "path",
                        "required": true,
                        "description": "The membrane ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "structure",
                        "in": "query",
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ],
                            "maxLength": 30
                        }
                    },
                    {
                        "name": "publication",
                        "in": "query",
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "minimum": 1
                        }
                    },
                    {
                        "name": "membrane",
                        "in": "query",
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "minimum": 1
                        }
                    },
                    {
                        "name": "method",
                        "in": "query",
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "minimum": 1
                        }
                    },
                    {
                        "name": "charge",
                        "in": "query",
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ],
                            "maxLength": 40
                        }
                    },
                    {
                        "name": "with_value",
                        "in": "query",
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ],
                            "enum": [
                                "logperm",
                                "logk",
                                "gpen",
                                "gwat",
                                "x_min",
                                null
                            ]
                        }
                    },
                    {
                        "name": "temperature_min",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "temperature_max",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "ph_min",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "ph_max",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "logperm_min",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "logperm_max",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "logk_min",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "logk_max",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "gpen_min",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "gpen_max",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "gwat_min",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "gwat_max",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "x_min_min",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "x_min_max",
                        "in": "query",
                        "schema": {
                            "type": [
                                "number",
                                "null"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "minimum": 1,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated set of `InteractionPassiveResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/InteractionPassiveResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object",
                                            "properties": {
                                                "first": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "last": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "prev": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "next": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "first",
                                                "last",
                                                "prev",
                                                "next"
                                            ]
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "from": {
                                                    "type": [
                                                        "integer",
                                                        "null"
                                                    ],
                                                    "minimum": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "description": "Generated paginator links.",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "url": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "label": {
                                                                "type": "string"
                                                            },
                                                            "active": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "required": [
                                                            "url",
                                                            "label",
                                                            "active"
                                                        ]
                                                    }
                                                },
                                                "path": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ],
                                                    "description": "Base path for paginator generated URLs."
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "description": "Number of items shown per page.",
                                                    "minimum": 0
                                                },
                                                "to": {
                                                    "type": [
                                                        "integer",
                                                        "null"
                                                    ],
                                                    "description": "Number of the last item in the slice.",
                                                    "minimum": 1
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "description": "Total number of items being paginated.",
                                                    "minimum": 0
                                                }
                                            },
                                            "required": [
                                                "current_page",
                                                "from",
                                                "last_page",
                                                "links",
                                                "path",
                                                "per_page",
                                                "to",
                                                "total"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "links",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EntityReferenceResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "abbreviation": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "abbreviation"
                ],
                "title": "EntityReferenceResource"
            },
...
            "InteractionPassiveResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "url": {
                        "type": "string"
                    },
                    "rdf": {
                        "type": "string",
                        "description": "IRI of the interaction in the MolMeDB RDF graph."
                    },
                    "structure": {
                        "$ref": "#/components/schemas/StructureReferenceResource"
                    },
                    "membrane": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/EntityReferenceResource"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "method": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/EntityReferenceResource"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "temperature": {
                        "type": [
                            "number",
                            "null"
                        ],
                        "description": "Temperature [°C]."
                    },
                    "ph": {
                        "type": [
                            "number",
                            "null"
                        ]
                    },
                    "charge": {
                        "type": "string",
                        "description": "Charge of the measured form as recorded, e.g. \"0\", \"+1\", \"-1\" or \"+1,-1\"."
                    },
                    "note": {
                        "type": "string"
                    },
                    "x_min": {
                        "type": [
                            "number",
                            "null"
                        ],
                        "description": "Xmin, the drug position with minimum energy in the membrane [nm]."
                    },
                    "x_min_accuracy": {
                        "type": [
                            "number",
                            "null"
                        ]
                    },
                    "gpen": {
                        "type": [
                            "number",
                            "null"
                        ],
                        "description": "ΔGpen, the penetration barrier [kcal/mol]."
                    },
                    "gpen_accuracy": {
                        "type": [
                            "number",
                            "null"
                        ]
                    },
                    "gwat": {
                        "type": [
                            "number",
                            "null"
                        ],
                        "description": "ΔGwat, the affinity towards the membrane [kcal/mol]."
                    },
                    "gwat_accuracy": {
                        "type": [
                            "number",
                            "null"
                        ]
                    },
                    "logk": {
                        "type": [
                            "number",
                            "null"
                        ],
                        "description": "LogK, the logarithm of the membrane-water partition coefficient [mol_m/mol_w]."
                    },
                    "logk_accuracy": {
                        "type": [
                            "number",
                            "null"
                        ]
                    },
                    "logperm": {
                        "type": [
                            "number",
                            "null"
                        ],
                        "description": "LogPerm, the logarithm of the permeability coefficient [cm/s]."
                    },
                    "logperm_accuracy": {
                        "type": [
                            "number",
                            "null"
                        ]
                    },
                    "primary_reference": {
                        "anyOf": [
                            {
                                "description": "Publication the value was taken from.",
                                "$ref": "#/components/schemas/PublicationReferenceResource"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "secondary_reference": {
                        "anyOf": [
                            {
                                "description": "Publication of the whole dataset (e.g. a review or a database the value was collected from).",
                                "$ref": "#/components/schemas/PublicationReferenceResource"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "required": [
                    "id",
                    "url",
                    "rdf",
                    "structure",
                    "membrane",
                    "method",
                    "temperature",
                    "ph",
                    "charge",
                    "note",
                    "x_min",
                    "x_min_accuracy",
                    "gpen",
                    "gpen_accuracy",
                    "gwat",
                    "gwat_accuracy",
                    "logk",
                    "logk_accuracy",
                    "logperm",
                    "logperm_accuracy",
                    "primary_reference",
                    "secondary_reference"
                ],
                "title": "InteractionPassiveResource"
            },
...
            "PublicationReferenceResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "pmid": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "citation": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "pmid",
                    "citation"
                ],
                "title": "PublicationReferenceResource"
            },
...
            "StructureReferenceResource": {
                "type": "object",
                "properties": {
                    "identifier": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    }
                },
                "required": [
                    "identifier",
                    "name",
                    "url"
                ],
                "title": "StructureReferenceResource"
            },
...
        },
        "responses": {
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            }
        }
    }
}